Open DAPR service listener even when no triggers present
Currently, we only open the DAPR app port when there are triggers
present. This means that different DAPR CLI commands are required when
there are and aren't triggers in the app which is not ideal. Therefore,
we always want to open this port.

To support this, we now no longer know the trigger routes at the time we
open the port, meaning we can't use the route builder. Instead, we have
a RequestDelegate that can handle dynamically added routes as we
register the proper triggers in the Functions pipeline.